Most playback functionality in place. Fixed bug where deleting playlist items wouldnt delete playlistIds. Fixed bug where overclicking an item for deletion would delete multiple items in the playlist

+++ app/src/main/java/com/gmail/markdevw/wetube/adapters/PlaylistItemAdapter.java
@@ -96,7 +96,9 @@ public class PlaylistItemAdapter extends RecyclerView.Adapter<PlaylistItemAdapte
         public void onClick(View view) {
             switch(view.getId()){
                 case R.id.playlist_delete:
-                    getDelegate().onDeleteItemClicked(PlaylistItemAdapter.this, playlistItem);
+                    if(!playlistItem.isToBeDeleted()){
+                        getDelegate().onDeleteItemClicked(PlaylistItemAdapter.this, playlistItem);
+                    }
                     break;
                 default:
                     getDelegate().onPlayListItemClicked(PlaylistItemAdapter.this, playlistItem, itemIndex, itemView);
